Fix defects detected by Coverity Scan

+++ src/core/receiver/gnss_flowgraph.cc
@@ -418,7 +418,7 @@ void GNSSFlowgraph::connect()
                                                 {
                                                     decimation--;
                                                 };
-                                            double acq_fs = fs / decimation;
+                                            double acq_fs = static_cast<double>(fs) / static_cast<double>(decimation);
 
                                             if (decimation > 1)
                                                 {
